[RN] Move features/app common functions to functions.common.js
