+++ react/features/app/functions.web.js
@@ -17,7 +17,6 @@ import { _configureLoggingLevels } from './functions.common';
 
 const Logger = require('jitsi-meet-logger');
 
-export { _parseURIString } from './functions.native';
 
 /**
  * Determines which route is to be rendered in order to depict a specific Redux
